requirements.txt
fastapi
uvicorn
strands-agents
python-dotenv
httpx
pydantic